+++ railties/lib/rails/generators/rails/controller/controller_generator.rb
@@ -15,12 +15,8 @@ module Rails
       end
 
       def add_routes
-        unless options[:skip_routes]
-          actions.reverse_each do |action|
-            # route prepends two spaces onto the front of the string that is passed, this corrects that.
-            route indent(generate_routing_code(action), 2)[2..-1]
-          end
-        end
+        return if options[:skip_routes]
+        route generate_routing_code(actions)
       end
 
       hook_for :template_engine, :test_framework, :helper, :assets
@@ -28,11 +24,12 @@ module Rails
       private
 
         # This method creates nested route entry for namespaced resources.
-        # For eg. rails g controller foo/bar/baz index
+        # For eg. rails g controller foo/bar/baz index show
         # Will generate -
         # namespace :foo do
         #   namespace :bar do
         #     get 'baz/index'
+        #     get 'baz/show'
         #   end
         # end
         def generate_routing_code(action)
@@ -49,7 +46,9 @@ module Rails
 
           # Create route
           #     get 'baz/index'
-          lines << indent(%{get '#{file_name}/#{action}'\n}, depth * 2)
+          actions.each do |action|
+            lines << indent(%{get '#{file_name}/#{action}'\n}, depth * 2)
+          end
 
           # Create `end` ladder
           #   end