+++ railties/CHANGELOG.md
@@ -1,3 +1,7 @@
+*   Optimize routes indentation.
+
+    *Yoshiyuki Hirano*
+
 *   Optimize indentation for generator actions.
 
     *Yoshiyuki Hirano*